This seems to speed things up for me pretty dramatically. Some very naive timing just with my stopwatch on my phone it goes like this:

main:
- Test output starting to show: 20 seconds
- Finished: 34 seconds

This branch:
- Test output starting to show: 6 seconds
- Finished: 20 seconds