@seblyng 2025-06-23
I see some tests are failing. I will look into them. I still didn't manage to get the tests running for some reason. I got past the error I talked about in the previous PR, but I just got hit with a bunch of other random errors.

Will maybe need to get them to work locally so I don't need to push to run the tests for this