@seblyng 2025-06-23
Hmmm, I finally got them to run locally somehow. They are passing for me🤔 I am using macos m2. Could this be a windows thing?